--- server/constructors.go
@@ -1,29 +1,0 @@
-package server
-
-import (
-	"io"
-	"os"
-	"path/filepath"
-
-	dbm "github.com/tendermint/tm-db"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-)
-
-func openDB(rootDir string) (dbm.DB, error) {
-	dataDir := filepath.Join(rootDir, "data")
-	db, err := sdk.NewLevelDB("application", dataDir)
-	return db, err
-}
-
-func openTraceWriter(traceWriterFile string) (w io.Writer, err error) {
-	if traceWriterFile != "" {
-		w, err = os.OpenFile(
-			traceWriterFile,
-			os.O_WRONLY|os.O_APPEND|os.O_CREATE,
-			0666,
-		)
-		return
-	}
-	return
-}
